Moving TimeString.h to basics.
