+++ arangod/Cluster/ActionBase.cpp
@@ -24,8 +24,8 @@
 
 #include "Cluster/ActionBase.h"
 
-#include "Agency/TimeString.h"
 #include "ApplicationFeatures/ApplicationServer.h"
+#include "Basics/TimeString.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/HeartbeatThread.h"
 #include "Cluster/MaintenanceFeature.h"
