+++ arangod/RestHandler/RestAdminClusterHandler.cpp
@@ -32,11 +32,11 @@
 
 #include "Agency/AgencyPaths.h"
 #include "Agency/AsyncAgencyComm.h"
-#include "Agency/TimeString.h"
 #include "Agency/TransactionBuilder.h"
 #include "ApplicationFeatures/ApplicationServer.h"
-#include "Basics/ResultT.h"
 #include "Basics/NumberUtils.h"
+#include "Basics/ResultT.h"
+#include "Basics/TimeString.h"
 #include "Cluster/AgencyCache.h"
 #include "Cluster/ClusterFeature.h"
 #include "Cluster/ClusterHelpers.h"
